requirement.txt
absl-py==0.15.0
cpbd==1.0.7
face-alignment==1.3.4
face-detection==0.2.2
Forward-Warp==0.0.1
forward-warp-cuda==0.0.0
h5py==3.1.0
imageio==2.15.0
keras==2.6.0
librosa==0.9.2
lpips==0.1.4
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
numpy-quaternion==2022.4.2
oauthlib==3.2.0
opencv-python==4.4.0.40
pandas==1.1.5
pfm==0.6.0
Pillow==8.4.0
protobuf==3.19.4
python-dateutil==2.8.2
python-speech-features==0.6
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.2.0
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorflow==2.6.2
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.64.1
xlrd==1.2.0